% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myparallel.R
\name{myparallel}
\alias{myparallel}
\title{Produce Parallel Coordinate Plot Imitating ggparcoord}
\usage{
myparallel(data, columns = 1:ncol(data), groupColumn = NULL,
  jitter = FALSE, jittervariables = NULL, jitterratio = 0.1,
  mapping = NULL, ...)
}
\arguments{
\item{data}{the data.frame to plot}

\item{columns}{a vector of variables (either names or indices) to be axes in the plot}

\item{groupColumn}{a single variable to group (color) by}

\item{jitter}{a single logical value, jitter the points of the selected varaible(s) or not}

\item{jittervariables}{varibles that will be jittered, should be a subset of columns, both indices or names;
if unspecified, all of the variables will be jittered}

\item{jitterratio}{the ratio of the maximum random ajustment devided by the range of selected variables (after scaling)}

\item{mapping}{NOT AVAILABLE NOW, aes string to pass to ggplot2 object}

\item{...}{NOT AVAILABLE NOW, passed on directly to all of the ggplot2 commands}
}
\value{
returns a ggplot2 object that can be plotted directly or used as base layers for additional modification
}
\description{
This function produce parallel coordinate plot pretty much like
GGally::ggparcoord. It supports color and jitter, and some
straight forward modifications from ggplot2. But some functionalities
are under work.
}
\examples{
# use sample of the diamonds data for illustrative purposes
data(diamonds, package="ggplot2")
diamonds.samp <- diamonds[sample(1:dim(diamonds)[1], 100), ]

# default use
myparallel(data = diamonds.samp, columns = c(1, 5:7))

# add color
myparallel(data = diamonds.samp, columns = c(1, 5:7), groupColumn = "cut")

# jitter and title
myparallel(data = diamonds.samp, columns = c(1, 5:7), groupColumn = "cut",
jitter = TRUE, jittervariables = "table") +
ggplot2::ggtitle("hello parallel")
}
